requirements.txt
flask==3.1.0
flask-socketio==5.5.0
python-socketio==5.12.1
python-engineio==4.11.2
eventlet==0.38.2
python-dotenv==1.0.1
anthropic==0.42.0
openai==1.57.0
google-generativeai==0.8.3
ptyprocess==0.7.0
pylama==8.4.1
werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
blinker==1.9.0
bidict==0.23.1
greenlet==3.1.1
dnspython==2.7.0
six==1.17.0
pywinpty==2.0.14